refactor for consistency of architecture CRUD access of data performed by model controller used to keep state and access model
