+++ lib/views/survey/survey_binding.dart
@@ -1,11 +1,11 @@
 import 'package:get/instance_manager.dart';
-import 'package:prapare/models/survey_model.dart';
+import 'package:prapare/controllers/questionnaire_controller.dart';
 import 'package:prapare/views/survey/survey_controller.dart';
 
 class SurveyBinding implements Bindings {
   @override
   void dependencies() {
     Get.put<SurveyController>(SurveyController());
-    Get.put<SurveyModel>(SurveyModel());
+    Get.put<QuestionnaireController>(QuestionnaireController());
   }
 }
